Add spellbook selection and add/remove buttons in the spell list.

+++ app/controllers/spellbooks_controller.rb
@@ -1,21 +1,26 @@
 class SpellbooksController < ApplicationController
+  before_action :logged_in_user, only: [:show, :new, :edit, :update, :create, :destroy, :spells, :inscribe, :erase, :select]
   autocomplete :spell, :name, full: true
 
   def index
-    @spellbooks = Spellbook.all
+    @spellbooks = policy_scope(Spellbook).all
   end
 
   def show
     @spellbook = Spellbook.find(params[:id])
+    authorize @spellbook
     @spells = @spellbook.spells
+    authorize @spells
   end
 
   def new
+    authorize Spellbook
     @spellbook = current_user.spellbooks.build
   end
 
   def create
     @spellbook = current_user.spellbooks.build(spellbook_params)
+    authorize @spellbook
     if @spellbook.save
       flash[:success] = 'Spellbook created!'
       redirect_to spellbooks_path
@@ -25,11 +30,13 @@ class SpellbooksController < ApplicationController
   end
 
   def edit
+    authorize Filter
     @spellbook = Spellbook.find(params[:id])
   end
 
   def update
     @spellbook = Spellbook.find(params[:id])
+    authorize @spellbook
     if @spellbook.update_attributes(spellbook_params)
       flash[:success] = 'Spellbook updated!'
       redirect_to spellbooks_path
@@ -40,6 +47,7 @@ class SpellbooksController < ApplicationController
 
   def destroy
     spellbook = Spellbook.find(params[:id])
+    authorize spellbook
     spellbook.destroy
     flash[:success] = 'Spellbook deleted!'
     redirect_to spellbooks_path
@@ -47,6 +55,7 @@ class SpellbooksController < ApplicationController
 
   def select
     spellbook = Spellbook.find(params[:id])
+    authorize spellbook
     select_spellbook(spellbook)
     flash[:success] = "Select Spellbook: #{current_spellbook.name}"
     redirect_to spells_path
@@ -55,6 +64,28 @@ class SpellbooksController < ApplicationController
   def spells
     @spellbook = Spellbook.find(params[:id])
     @spells = @spellbook.spells
+  end
+
+  def inscribe
+    authorize current_spellbook
+
+    @spell = Spell.find(params[:spell_id])
+    current_spellbook.inscribe(@spell)
+    respond_to do |format|
+      format.html { redirect_to @spell}
+      format.js
+    end
+  end
+
+  def erase
+    authorize current_spellbook
+
+    @spell = Spell.find(params[:spell_id])
+    current_spellbook.erase(@spell)
+    respond_to do |format|
+      format.html { redirect_to @spell}
+      format.js
+    end
   end
 
   private
